+++ CHANGELOG.md
@@ -18,6 +18,7 @@ außerdem befolgt dieses Projekt die [Semantische Versionierung](https://semver.
   - bei nicht erfolgreichen Verbindungen wird der Button wieder deaktiviert
 - Modbus Adresse wird beim Auslesen der Schreib.-/ Lese-Register aktualisiert
 - Haken MCS Konfig wird automatisch gesetzt wenn die Platine entsprechend konfiguriert wurde
+- Hilfe ist aus dem Menu aufrufbar
 
 ### Geändert